Meeple bid action에 따라 owner 변경 처리
