+++ server/turn.go
@@ -2,6 +2,7 @@ package server
 
 import (
 	"Server/enum"
+	"Server/game"
 	"encoding/json"
 	"log"
 )
@@ -13,7 +14,6 @@ func (hub *Hub) nextTurn() {
 	hub.currentTurn++
 	if hub.currentTurn > hub.PlayerNum {
 		hub.currentTurn = 1
-		hub.playerSkipCnt = 0
 	}
 }
 
@@ -21,6 +21,29 @@ func (hub *Hub) sendTurnChangeData() {
 	data, _ := json.Marshal(TurnChangeData{Turn: hub.currentTurn})
 	message := Message{Type: enum.TurnChange, Data: string(data)}
 	hub.broadcast(&message)
+}
+
+func (hub *Hub) nextSeason() {
+	hub.mu.Lock()
+	defer hub.mu.Unlock()
+
+	switch hub.currentSeason {
+	case enum.Spring:
+		tiles := game.GetRoundTilesBySeason(enum.Summer, 6)
+		for _, tile := range tiles {
+			hub.putTile(tile)
+		}
+		hub.currentSeason = enum.Summer
+	case enum.Summer:
+		hub.currentSeason = enum.Autumn
+	case enum.Autumn:
+		hub.currentSeason = enum.Winter
+	case enum.Winter:
+		// TODO 겨울 라운드가 끝나고 점수 계산하는 기능 추가
+		break
+	}
+	hub.playerSkipCnt = 0
+	hub.sendSeasonChangeData()
 }
 
 func (hub *Hub) sendSeasonChangeData() {
